+++ lib/src/menu/widgets/button.dart
@@ -1,11 +1,15 @@
 /// Provides the [Button] class.
+import '../../json/sound_reference.dart';
 import 'widgets_base.dart';
 
 /// A button that can be activated.
 class Button extends Widget {
   /// Create a button.
-  Button(this.onActivate);
+  Button(this.onActivate, {this.sound});
 
   /// The function which will be called when this button is activated.
   final void Function() onActivate;
+
+  /// The sound that should play when this widget is activated.
+  final SoundReference? sound;
 }